refactor(chat): split chatx reducer
