+++ src/reducers/initialState.js
@@ -9,11 +9,4 @@ export default {
     initialized: false,
   },
   ajaxCallsInProgress: 0,
-  chatx: {
-    enabled: false,
-    users: {},
-    rooms: {},
-    joinedRooms: {},
-    currentRoom: null,
-  },
 };
